requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djoser==2.1.0
Pillow==9.2.0
drf-yasg==1.21.4
